Reorganize uri-template into per-feature module layouts
Splits each feature module into a thin entry plus dedicated
implementation file:

 -  src/template/mod.ts now re-exports from template.ts and errors.ts.
    Template class implementation moves to src/template/template.ts.
 -  src/router/mod.ts now re-exports from router.ts and errors.ts.
    Router class implementation moves to src/router/router.ts.

Consolidates parse/expansion errors with the template feature:

 -  src/errors.ts moves to src/template/errors.ts with internal
    callers updated.

Reorganizes test fixtures and helpers per feature:

 -  src/tests/lib.ts moves to src/tests/template.ts.
 -  src/tests/json/{hard,match,wrong}.json move under
    src/tests/json/template/.

All imports across the package are updated to the new locations.
