+++ packages/uri-template/src/mod.ts
@@ -7,6 +7,12 @@
  * @module
  */
 
+export { Router, RouterError, RouteTemplatePathError } from "./router/mod.ts";
+export type {
+  RouterOptions,
+  RouterPathPattern,
+  RouterRouteResult,
+} from "./router/router.ts";
 export {
   EmptyExpressionError,
   EmptyVarNameError,
@@ -18,21 +24,14 @@ export {
   PrefixModifierNotApplicableError,
   ReservedOperatorError,
   StrayClosingBraceError,
+  Template,
   TemplateExpansionError,
   TemplateParseError,
   TrailingCommaError,
   UnclosedExpressionError,
   UnexpectedCharacterError,
   UnknownOperatorError,
-} from "./errors.ts";
-export { RouterError, RouteTemplatePathError } from "./router/errors.ts";
-export { default as Router } from "./router/mod.ts";
-export type {
-  RouterOptions,
-  RouterPathPattern,
-  RouterRouteResult,
-} from "./router/mod.ts";
-export { default as Template } from "./template/mod.ts";
+} from "./template/mod.ts";
 export type {
   AssociativeValue,
   ExpandContext,
